backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
openai
PyGithub
redis
supabase

pytest
httpx
langchain-openai
tree-sitter-languages
